offline motor will be disabled for 6 seconds

+++ driver/inc/Driver_Motor.h
@@ -24,7 +24,7 @@ typedef struct {
     int16_t  speed;         // 转子转速值, RPM
     float    actualCurrent; //电流
     int16_t  temperature;   //温度
-    uint16_t updatedAt;     // 更新电机数据时的tickCount
+    uint32_t updatedAt;     // 更新电机数据时的tickCount
     uint8_t  online;        // 电机是否离线
 
     int8_t  angleEnabled;  // 连续角度启用
